requirements.txt
# AI Urban Warfare Analyst - Python Dependencies

# Core ML/AI
google-generativeai>=0.5.0
openai>=1.0.0

# Video Processing
opencv-python>=4.8.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Async Operations (for Stage 2)
aiohttp>=3.9.0
nest-asyncio>=1.6.0

# Environment Management
python-dotenv>=1.0.0

# Jupyter Support
ipython>=8.0.0
ipywidgets>=8.0.0
jupyter>=1.0.0

# Optional: Progress Bars
tqdm>=4.65.0
